+++ src/main/java/game/board/entities/playerentities/soldiers/FastSoldier.java
@@ -18,7 +18,7 @@ public class FastSoldier extends Soldier {
      * @param numberOfTileHeCanJump
      * The number of tile (step) the fast soldier can move per time
      */
-    public FastSoldier(Position position, Player owner, double price, int healthPoint, int numberOfMoveAtEachRound, int numberOfTileHeCanJump) {
+    public FastSoldier(Position position, Player owner, int price, int healthPoint, int numberOfMoveAtEachRound, int numberOfTileHeCanJump) {
         super(position, owner, price, healthPoint, numberOfMoveAtEachRound);
         this.numberOfTileHeCanJump = numberOfTileHeCanJump;
     }
